refactor(neuron-ui): change Header, Sidebar, Notification, MainContent to container
