+++ packages/neuron-ui/src/containers/MainContent/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const MainContent = ({ children }: { children?: React.ReactNode }) => <main>{children}</main>
+
+MainContent.displayName = 'MainContent'
+
+export default MainContent
